# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Simple EC Backend API
  description: |
    A simple e-commerce backend for comparing Stream/Sequence performance in 1-to-many data export.
    This project is an experimental repository for Advent Calendar article.
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /admin/seed:
    post:
      tags:
        - admin
      summary: Seed test data
      description: Generate test data for development/testing. Use seed parameter for reproducible data generation.
      operationId: seedData
      parameters:
        - name: customers
          in: query
          description: Number of customers to generate
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 10000
        - name: orders
          in: query
          description: Number of orders to generate
          schema:
            type: integer
            default: 1000
            minimum: 1
            maximum: 100000
        - name: attrs
          in: query
          description: Number of attribute definitions to generate (0 = no attributes)
          schema:
            type: integer
            default: 0
            minimum: 0
            maximum: 100
        - name: seed
          in: query
          description: Random seed for reproducible data generation (optional, random if not specified)
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeedDataResponse'
...
components:
  schemas:
...
    SeedDataResponse:
      type: object
      required:
        - customersCreated
        - ordersCreated
        - orderItemsCreated
        - attributeDefinitionsCreated
        - attributeValuesCreated
      properties:
        customersCreated:
          type: integer
          description: Number of customers created
        ordersCreated:
          type: integer
          description: Number of orders created
        orderItemsCreated:
          type: integer
          description: Number of order items created
        attributeDefinitionsCreated:
          type: integer
          description: Number of attribute definitions created
        attributeValuesCreated:
          type: integer
          description: Number of attribute values created
...